Successfully extracted brain.
