+++ hangry_games/src/tributes/actors.rs
@@ -2,8 +2,9 @@ use std::str::FromStr;
 use crate::areas::Area;
 use rand::prelude::*;
 
-use super::actions::{TributeAction, AttackResult, AttackOutcome};
+use super::actions::{AttackOutcome, AttackResult};
 use super::statuses::TributeStatus;
+use super::brains::TributeBrain;
 
 
 #[derive(Clone, Debug, PartialEq)]
@@ -35,136 +36,6 @@ pub struct Tribute {
     pub is_hidden: Option<bool>,
     pub dexterity: Option<i32>,
     pub status: TributeStatus
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct TributeBrain {
-    previous_actions: Vec<TributeAction>,
-    preferred_action: Option<TributeAction>,
-    preferred_action_percentage: f64,
-}
-
-impl TributeBrain {
-    fn new() -> Self {
-        Self {
-            previous_actions: Vec::new(),
-            preferred_action: None,
-            preferred_action_percentage: 0.0,
-        }
-    }
-
-    pub fn set_preferred_action(&mut self, action: TributeAction, percentage: f64) {
-        self.preferred_action = Some(action);
-        self.preferred_action_percentage = percentage;
-    }
-
-    pub fn clear_preferred_action(&mut self) {
-        self.preferred_action = None;
-        self.preferred_action_percentage = 0.0;
-    }
-
-    /// Decide on an action for the tribute to take
-    /// First weighs any preferred actions, then decides based on current state
-    pub fn act(&mut self, tribute: &Tribute, nearby_tributes: Vec<Tribute>) -> TributeAction {
-        if tribute.health == 0 { return TributeAction::None; }
-
-        if let Some(preferred_action) = self.clone().preferred_action {
-            if thread_rng().gen_bool(self.preferred_action_percentage) {
-                self.previous_actions.push(preferred_action.clone());
-                return preferred_action;
-            }
-        }
-
-        let action = TributeBrain::decide_on_action(tribute, nearby_tributes.clone());
-
-        // Try to get a different action?
-
-        self.previous_actions.push(action.clone());
-        action
-    }
-
-    /// Get the last action taken by the tribute
-    pub fn last_action(&self) -> TributeAction {
-        if let Some(previous_action) = self.previous_actions.last() {
-            previous_action.clone()
-        } else {
-            TributeAction::None
-        }
-    }
-
-    /// The AI for a tribute. Automatic decisions based on current state.
-    fn decide_on_action(tribute: &Tribute, nearby_tributes: Vec<Tribute>) -> TributeAction {
-        // If the tribute isn't in the area, they do nothing
-        if tribute.area.is_none() {
-            return TributeAction::None;
-        }
-        if tribute.movement <= 0 {
-            return TributeAction::Rest;
-        }
-
-        let _area = tribute.area.as_ref().unwrap();
-
-        match &nearby_tributes.len() {
-            0 => {
-                match tribute.health {
-                    // health is low, rest
-                    1..=20 => TributeAction::Rest,
-                    // health isn't great, hide
-                    // unless sanity is also low, then move
-                    21..=30 => {
-                        if tribute.sanity > 20 && tribute.is_visible() {
-                            TributeAction::Hide
-                        } else {
-                            TributeAction::Move
-                        }
-                    },
-                    // health is good, move
-                    _ => {
-                        // If the tribute has movement, move
-                        match tribute.movement {
-                            0 => TributeAction::Rest,
-                            _ => TributeAction::Move,
-                        }
-                    }
-                }
-            }
-            1..6 => {
-                // Enemies are nearby, attack depending on health
-                match tribute.health {
-                    // health is low, hide
-                    1..=5 => {
-                        if tribute.sanity > 20 && tribute.is_visible() {
-                            TributeAction::Hide
-                        } else {
-                            TributeAction::Attack
-                        }
-                    },
-                    // health isn't great, run away
-                    6..=10 => {
-                        if tribute.sanity > 20 {
-                            TributeAction::Move
-                        } else {
-                            TributeAction::Attack
-                        }
-                    },
-                    // health is good, attack
-                    _ => TributeAction::Attack,
-                }
-            },
-            _ => {
-                // More than 5 enemies? Intelligence decides next move
-                let sense = 100 - tribute.intelligence.unwrap() - tribute.sanity;
-                match sense {
-                    // Too dumb to know better, attacks
-                    0..36 => TributeAction::Attack,
-                    // Smart enough to know better, hides
-                    85..101 => TributeAction::Hide,
-                    // Average intelligence, moves
-                    _ => TributeAction::Move,
-                }
-            }
-        }
-    }
 }
 
 
@@ -208,6 +79,7 @@ impl Tribute {
     /// Reduces health
     pub fn takes_physical_damage(&mut self, damage: i32) {
         self.health = std::cmp::max(0, self.health - damage);
+        self.status = TributeStatus::Wounded;
 
         if self.health == 0 {
             self.dies();
@@ -480,6 +352,7 @@ impl From<TributeModel> for Tribute {
 }
 
 use crate::models::tribute::UpdateTribute;
+
 impl Into<UpdateTribute> for Tribute {
     fn into(self) -> UpdateTribute {
         let area = self.area.as_ref().unwrap();
@@ -540,9 +413,8 @@ mod tests {
     #[test]
     fn moves_and_rests() {
         let mut tribute = Tribute::new("Katniss".to_string(), None);
-        tribute.speed = Some(10);
         tribute.moves();
-        assert_eq!(tribute.movement, 90);
+        assert_eq!(tribute.movement, 50);
         tribute.rests();
         assert_eq!(tribute.movement, 100);
     }
@@ -551,54 +423,7 @@ mod tests {
     fn takes_damage_and_dies() {
         let mut tribute = Tribute::new("Katniss".to_string(), None);
         tribute.takes_physical_damage(100);
-        assert_eq!(tribute.status, TributeStatus::Dead);
-    }
-
-    #[test]
-    fn decide_on_action_default() {
-        // If there are no enemies nearby, the tribute should move
-        let mut tribute = Tribute::new("Katniss".to_string(), None);
-        let action = tribute.brain.act(&tribute.clone(), vec![]);
-        assert_eq!(action, TributeAction::Move);
-    }
-
-    #[test]
-    fn decide_on_action_low_health() {
-        // If the tribute has low health, they should rest
-        let mut tribute = Tribute::new("Katniss".to_string(), None);
-        tribute.takes_physical_damage(90);
-        let action = tribute.brain.act(&tribute.clone(), vec![]);
-        assert_eq!(action, TributeAction::Rest);
-    }
-
-    #[test]
-    fn decide_on_action_no_movement() {
-        // If the tribute has no movement, they should rest
-        let mut tribute = Tribute::new("Katniss".to_string(), None);
-        tribute.speed = Some(100);
-        tribute.moves();
-        let action = tribute.brain.act(&tribute.clone(), vec![]);
-        assert_eq!(action, TributeAction::Rest);
-    }
-
-    #[test]
-    fn decide_on_action_enemies() {
-        // If there are enemies nearby, the tribute should attack
-        let mut tribute = Tribute::new("Katniss".to_string(), None);
-        let tribute2 = Tribute::new("Peeta".to_string(), None);
-        let action = tribute.brain.act(&tribute.clone(), vec![tribute.clone(), tribute2]);
-        assert_eq!(action, TributeAction::Attack);
-    }
-
-    #[test]
-    fn decide_on_action_enemies_low_health() {
-        // If there are enemies nearby, but the tribute is low on health
-        // the tribute should hide
-        let mut tribute = Tribute::new("Katniss".to_string(), None);
-        tribute.takes_physical_damage(90);
-        let tribute2 = Tribute::new("Peeta".to_string(), None);
-        let action = tribute.brain.act(&tribute.clone(),vec![tribute.clone(), tribute2]);
-        assert_eq!(action, TributeAction::Hide);
+        assert_eq!(tribute.status, TributeStatus::RecentlyDead);
     }
 
     #[test]
