+++ hangry_games/src/tributes/mod.rs
@@ -1,3 +1,4 @@
 pub mod actors;
 pub mod actions;
 pub mod statuses;
+pub mod brains;
